Fix NGINX Security Hole Properly
Accessing the CA from a shared namespace outside of the current release
is banned (by default) by the NGINX ingress controller to prevent it
from being able to read every secret on the entire cluster.  You can
enable this broken behaviour with `allow-cross-namespace-resources` but
being the professionals we are we keep the safe defaults and force every
mTLS enabled service to create a client certificate that cert-manager
will helpfully populate with a `ca.crt` key.